+++ src/gui/components/keybinds.rs
@@ -204,13 +204,7 @@ where
                         r
                     });
                     changed |= r.changed();
-
-                    ui.allocate_space(egui::vec2(1.0, 200.0));
                 });
-
-                if ui.available_height() > 0.0 {
-                    ui.allocate_space(ui.available_size());
-                }
             })
             .response
         });
@@ -250,6 +244,10 @@ impl egui::Widget for CommandSelectWidget<'_, GlobalKeybindsAccessor> {
                     "Save" => Cmd::Save,
                     "Save as..." => Cmd::SaveAs,
                     "Exit" => Cmd::Exit,
+
+                    "Copy .hsc" => Cmd::CopyHscLog,
+                    "Copy .log" => Cmd::CopyMc4dLog,
+                    "Paste .log" => Cmd::PasteLog,
 
                     "Undo" => Cmd::Undo,
                     "Redo" => Cmd::Redo,
@@ -453,7 +451,7 @@ pub trait KeybindSetAccessor: 'static + Clone + Hash + Send + Sync {
 
     fn get<'a>(&self, prefs: &'a Preferences) -> &'a [Keybind<Self::Command>];
     fn get_mut<'a>(&self, prefs: &'a mut Preferences) -> &'a mut Vec<Keybind<Self::Command>>;
-    fn get_defaults<'a>(&self) -> &'static [Keybind<Self::Command>] {
+    fn get_defaults(&self) -> &'static [Keybind<Self::Command>] {
         self.get(&crate::preferences::DEFAULT_PREFS)
     }