+++ src/gui/windows/keybinds_reference.rs
@@ -91,7 +91,7 @@ fn build(ui: &mut egui::Ui, app: &mut App) {
         prefs_ui.checkbox("Function keys", access!(.function));
         prefs_ui.checkbox("Navigation keys", access!(.navigation));
         prefs_ui.checkbox("Numpad", access!(.numpad));
-        prefs_ui.float("Max font size", access!(.max_font_size), |dv| {
+        prefs_ui.num("Max font size", access!(.max_font_size), |dv| {
             dv.fixed_decimals(1).clamp_range(1.0..=3.0).speed(0.01)
         });
 
@@ -244,6 +244,10 @@ fn draw_key(ui: &mut egui::Ui, app: &mut App, key: KeyMappingCode, rect: egui::R
                 Command::SaveAs => ui.label("Save As"),
                 Command::Exit => ui.label("Exit"),
 
+                Command::CopyHscLog => ui.label("Copy puzzle log (.hsc)"),
+                Command::CopyMc4dLog => ui.label("Copy puzzle log (.log)"),
+                Command::PasteLog => ui.label("Paste puzzle log"),
+
                 Command::Undo => ui.label("Undo"),
                 Command::Redo => ui.label("Redo"),
                 Command::Reset => ui.label("Reset"),
@@ -287,7 +291,7 @@ fn autosize_button_text(
         }
         font_size -= 1.0;
     }
-    egui::RichText::new("")
+    egui::RichText::new(".")
 }
 
 macro_rules! keyboard_key {